refactor: improve native tests
