+++ packages/video/__tests__/android/video.android.test.js
@@ -1,7 +1,3 @@
-import shared from "../shared";
-import sharedNative from "../shared-native";
+import shared from "../video-shared-native";
 
-describe("Video on native", () => {
-  shared();
-  sharedNative();
-});
+shared();
