+++ packages/video/__tests__/video-shared-native.js
@@ -1,10 +1,33 @@
 import React from "react";
 import { shallow } from "enzyme";
-
+import {
+  addSerializers,
+  compose,
+  enzymeRenderedSerializer,
+  flattenStyleTransform,
+  minimalNativeTransform,
+  print
+} from "@times-components/jest-serializer";
 import Video from "../src/video";
-import { defaultVideoProps } from "./shared";
+import defaultVideoProps from "./default-video-props";
 
 export default () => {
+  addSerializers(
+    expect,
+    enzymeRenderedSerializer(),
+    compose(print, flattenStyleTransform, minimalNativeTransform)
+  );
+
+  it("1. video", () => {
+    const wrapper = shallow(<Video {...defaultVideoProps} />);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it("2. video without a poster image", () => {
+    const wrapper = shallow(<Video {...defaultVideoProps} poster={null} />);
+    expect(wrapper).toMatchSnapshot();
+  });
+
   it("calls onVideoPress when the component is pressed", () => {
     const onVideoPress = jest.fn();
     const component = shallow(
