+++ packages/video/__tests__/web/inline-video-player.web.test.js
@@ -1,9 +1,8 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import { shallow } from "enzyme";
-
-import { defaultVideoProps } from "../shared";
 import InlineVideoPlayer from "../../src/inline-video-player.web";
+import defaultVideoProps from "../default-video-props";
 
 describe("InlineVideoPlayer", () => {
   afterEach(() => {
