+++ packages/video/__tests__/web/video.web.test.js
@@ -1,31 +1,20 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import videoProps from "../default-video-props";
-import Video from "../../src/video";
-
 import {
   addSerializers,
   compose,
   minimaliseTransform,
   print
 } from "@times-components/jest-serializer";
+import Video from "../../src/video";
+import defaultVideoProps from "../default-video-props";
 
 addSerializers(
   expect,
-  compose(
-    print,
-    minimaliseTransform((value, key) => key === "style")
-  )
+  compose(print, minimaliseTransform((value, key) => key === "style"))
 );
 
 it("1. video", () => {
-  const tree = renderer.create(<Video {...videoProps} />);
-  expect(tree).toMatchSnapshot();
-});
-
-it("2. video without a poster image", () => {
-  const tree = renderer.create(
-    <Video {...videoProps} poster={null} />
-  );
+  const tree = renderer.create(<Video {...defaultVideoProps} />);
   expect(tree).toMatchSnapshot();
 });
