+++ packages/video/src/is-paid-subscriber.web.js
